requirements.txt
django
djangorestframework
pdfkit
reportlab
gunicorn
djangorestframework-simplejwt
django-cors-headers
whitenoise
django-filter